ensure 328PB interrupt vectors are linked into sketch
Changing the file from a .h file to a .cpp file makes Arduino compile and link
it into the sketch automatically. This should be more reliable than the old
trick of including it from pins_arduino.h only when ARDUINO_MAIN is defined,
which effectively added it only to wiring_digital.c (to avoid multiple
definitions), but it was possible for a sketch to not need anything from
wiring_digital and for the new vectors to be omitted.